refactor(mentor): 새로 생성한 utils 디렉토리로 이메일 유효성 검사 파일을 옮기고 파일명 소문자로 수정
